Revert unnecessary action version bump.
Actions are supposed to update their "vN" tags to the latest minor version; e.g. checkout@v2 is the same as checkout@v2.3.4, but when 2.3.5 is released, then checkout@v2 will be changed to be the same as checkout@v2.3.5. It is a separate question whether we should use specific versions for stable builds, but this is not something that dependabot should decide for us.